Reimplement 'description' filter using DescriptionField
